# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API E-Commerce UAS",
        "description": "Dokumentasi lengkap API untuk integrasi Frontend Mobile (Flutter).",
        "contact": {
            "email": "hifadlibilal@sistekin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/locations": {
            "get": {
                "tags": [
                    "RajaOngkir"
                ],
                "summary": "Cari Lokasi (Autocomplete)",
                "description": "Endpoint untuk fitur pencarian alamat di Frontend. Ketik nama kecamatan/kota, sistem akan mengembalikan list lokasi beserta ID-nya.",
                "operationId": "c5cb78f8a74ec44be7aa32ef0c88f24e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Minimal 3 karakter. Contoh: 'Gubeng'",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Gubeng"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data lokasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "SIMPAN ID INI untuk Register",
                                                "type": "integer",
                                                "example": 69335
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "TAMBAKSARI, SURABAYA, JAWA TIMUR..."
                                            },
                                            "city_name": {
                                                "type": "string",
                                                "example": "SURABAYA"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi Gagal (Search kurang dari 3 huruf)"
                    }
                }
            }
        },
        "/check-ongkir": {
            "post": {
                "tags": [
                    "RajaOngkir"
                ],
                "summary": "Cek Biaya Ongkir (Untuk Menu Pilihan Kurir)",
                "description": "Frontend mengirim Origin (Lokasi Seller), Destination (Lokasi Buyer), Berat, dan Kurir. API akan mengembalikan daftar layanan (REG, BEST, YES) beserta harganya.",
                "operationId": "824c45374773dc69cebd988d456a8e6c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "description": "ID Lokasi Pengirim (Seller)",
                                        "type": "integer",
                                        "example": 444
                                    },
                                    "destination": {
                                        "description": "ID Lokasi Penerima (User yang sedang login)",
                                        "type": "integer",
                                        "example": 69335
                                    },
                                    "weight": {
                                        "description": "Berat dalam Gram",
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "courier": {
                                        "description": "Kode kurir: jne, sicepat, pos, jnt",
                                        "type": "string",
                                        "example": "jne"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List opsi pengiriman berhasil didapat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer di sini untuk mengakses endpoint yang dilindungi.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}